backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.23
requests==2.31.0
beautifulsoup4==4.12.2
korean-lunar-calendar==0.3.1
pandas==2.1.3
numpy==1.24.3
tensorflow==2.15.0
scikit-learn==1.3.2
pydantic-settings==2.1.0
email-validator==2.1.0
python-multipart==0.0.6
python-dateutil==2.8.2
yt-dlp==2024.7.25
google-api-python-client==2.139.0
youtube-transcript-api==0.6.2
openai==1.57.0
tiktoken==0.8.0
langchain==0.3.14
sentence-transformers==3.3.1
whisper==1.1.10
moviepy==1.0.3
pydub==0.25.1
speechrecognition==3.11.0
torch==2.5.1
torchaudio==2.5.1